completed my question3 of section 'Fundamental Data Types'

+++ functions/getDistanceFallen.cpp
@@ -1,6 +1,6 @@
 #include "getDistanceFallen.h"
 
-double getDistanceFallen(double seconds)
+double getDistanceFallen(int seconds)
 {
     return ( (9.8 * (seconds * seconds)) / 2 );
 }
