+++ headers/getBallPosition.h
@@ -4,6 +4,6 @@
 // Function to calculate the position of the ball
 // relative to the tower height
 // learn more in the files on the "practice-resources" folder
-double getBallPosition(double towerHeight, double seconds);
+double getBallPosition(double towerHeight, int seconds);
 
 #endif // GET_BALL_POSITION
