+++ headers/getDistanceFallen.h
@@ -3,6 +3,6 @@
 
 // Function to calculate distance fallen by a thrown ball
 // learn more in the files on the "practice-resources" folder
-double getDistanceFallen(double seconds);
+double getDistanceFallen(int seconds);
 
 #endif // GET_DISTANCE_FALLEN
